# services/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeTourist API",
    "description": "Backend for DeTourist",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trips/{trip_id}/visits": {
      "post": {
        "tags": [
          "visits"
        ],
        "summary": "Create Visit",
        "operationId": "create_visit_api_v1_trips__trip_id__visits_post",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trips/{trip_id}/visits/bulk": {
      "post": {
        "tags": [
          "visits"
        ],
        "summary": "Create Visits Bulk",
        "operationId": "create_visits_bulk_api_v1_trips__trip_id__visits_bulk_post",
        "security": [
          {
            "JWT": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitBulkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VisitResponse"
                  },
                  "title": "Response Create Visits Bulk Api V1 Trips  Trip Id  Visits Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisitBulkCreate": {
        "properties": {
          "visits": {
            "items": {
              "$ref": "#/components/schemas/VisitCreate"
            },
            "type": "array",
            "title": "Visits"
          }
        },
        "type": "object",
        "required": [
          "visits"
        ],
        "title": "VisitBulkCreate"
      },
      "VisitCreate": {
        "properties": {
          "poi_id": {
            "type": "string",
            "format": "uuid",
            "title": "Poi Id"
          },
          "arrived_at": {
            "type": "string",
            "format": "date-time",
            "title": "Arrived At"
          },
          "source": {
            "$ref": "#/components/schemas/VisitSource",
            "default": "gps_auto"
          },
          "departed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departed At"
          }
        },
        "type": "object",
        "required": [
          "poi_id",
          "arrived_at"
        ],
        "title": "VisitCreate"
      },
      "VisitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trip Id"
          },
          "poi_id": {
            "type": "string",
            "format": "uuid",
            "title": "Poi Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "arrived_at": {
            "type": "string",
            "format": "date-time",
            "title": "Arrived At"
          },
          "departed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departed At"
          },
          "source": {
            "$ref": "#/components/schemas/VisitSource"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "poi_id",
          "user_id",
          "arrived_at",
          "departed_at",
          "source",
          "created_at"
        ],
        "title": "VisitResponse"
      },
      "VisitSource": {
        "type": "string",
        "enum": [
          "gps_auto",
          "manual_checkin"
        ],
        "title": "VisitSource"
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}